ref(onboarding): Split console onboarding docs into multiple files
