+++ static/app/gettingStartedDocs/console/nintendo-switch/index.tsx
@@ -1,0 +1,8 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {onboarding} from 'sentry/gettingStartedDocs/console/nintendo-switch/onboarding';
+
+const docs: Docs = {
+  onboarding,
+};
+
+export default docs;
