+++ static/app/gettingStartedDocs/console/playstation/index.tsx
@@ -1,0 +1,10 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {onboarding} from 'sentry/gettingStartedDocs/console/playstation/onboarding';
+import {platformOptions} from 'sentry/gettingStartedDocs/console/playstation/utils';
+
+const docs: Docs = {
+  onboarding,
+  platformOptions,
+};
+
+export default docs;
